[ALPS04715419] Platform: 64bit project build error fix
Kernel fix various build errors for 64 bit project

MTK-Commit-Id: d52f9a4d5d802df3e7ecd4d5fcd164a285fdaf99

CR-Id: ALPS04715419
Feature: Configuration

+++ drivers/misc/mediatek/base/power/mt6757/mtk_udi.c
@@ -309,7 +309,7 @@ out0:
 static int udi_reg_proc_show(struct seq_file *m, void *v)
 {
 	seq_printf(m, "Reg[%x] = 0x%zx.\n", udi_addr_phy,
-		   udi_reg_read(udi_addr_phy));
+		   (size_t)udi_reg_read(udi_addr_phy));
 	return 0;
 }
 
@@ -335,17 +335,17 @@ static ssize_t udi_reg_proc_write(struct file *file, const char __user *buffer,
 		udi_reg_field(udi_addr_phy, udi_reg_msb
 			      : udi_reg_lsb, udi_value);
 		udi_info("Read back, Reg[%x] = 0x%zx.\n", udi_addr_phy,
-			 udi_reg_read(udi_addr_phy));
+			 (size_t)udi_reg_read(udi_addr_phy));
 	} else if (sscanf(buf, "%1s %x %x", udi_rw, &udi_addr_phy,
 			  &udi_value) == 3) {
 		/* w format or 'w', addr, value */
 		udi_reg_write(udi_addr_phy, udi_value);
 		udi_info("Read back, Reg[%x] = 0x%zx.\n", udi_addr_phy,
-			 udi_reg_read(udi_addr_phy));
+			 (size_t)udi_reg_read(udi_addr_phy));
 	} else if (sscanf(buf, "%1s %x", udi_rw, &udi_addr_phy) == 2) {
 		/* r format or 'r', addr */
 		udi_info("Read back, aReg[%x] = 0x%zx.\n", udi_addr_phy,
-			 udi_reg_read(udi_addr_phy));
+			 (size_t)udi_reg_read(udi_addr_phy));
 	} else {
 		udi_info("echo dbg_lv (dec) > /proc/udi/udi_debug\n");
 		memset(udi_rw, 0, sizeof(udi_rw));
